backend/requirements.txt
fastapi==0.95.2
uvicorn
requests
python-dotenv
pydantic==1.10.15
python-multipart
langdetect 
googletrans==4.0.0-rc1